+++ gitpush/__init__.py
@@ -5,6 +5,8 @@ One Command To Rule Them All
 Created by: Himanshu Kumar (@himanshu231204)
 """
 
-__version__ = "1.0.3"
+__version__ = "1.0.4"
+__license__ = "MIT"
+__copyright__ = "Copyright 2026 Himanshu Kumar"
 __author__ = "Himanshu Kumar"
 __github__ = "@himanshu231204"